refactor: polish control-plane store implementation
