+++ internal/persis/controlplanestore/postgres/integration_test.go
@@ -25,6 +25,7 @@ import (
 	"github.com/dagucloud/dagu/internal/service/audit"
 	"github.com/dagucloud/dagu/internal/service/eventstore"
 	"github.com/dagucloud/dagu/internal/workspace"
+	coordinatorv1 "github.com/dagucloud/dagu/proto/coordinator/v1"
 )
 
 func TestPostgresControlPlaneStoreIntegration(t *testing.T) {
@@ -116,17 +117,32 @@ func TestPostgresControlPlaneStoreIntegration(t *testing.T) {
 		gotWorkspace, err := store.Workspaces().GetByName(ctx, "ops")
 		require.NoError(t, err)
 		assert.Equal(t, ws.ID, gotWorkspace.ID)
+		ws.Description = "Ops Updated"
+		require.NoError(t, store.Workspaces().Update(ctx, ws))
+		updatedWorkspace, err := store.Workspaces().GetByID(ctx, ws.ID)
+		require.NoError(t, err)
+		assert.Equal(t, "Ops Updated", updatedWorkspace.Description)
 
 		user := auth.NewUser("admin", "hash", auth.RoleAdmin)
 		require.NoError(t, store.Users().Create(ctx, user))
 		gotUser, err := store.Users().GetByUsername(ctx, "admin")
 		require.NoError(t, err)
 		assert.Equal(t, user.ID, gotUser.ID)
+		user.IsDisabled = true
+		require.NoError(t, store.Users().Update(ctx, user))
+		updatedUser, err := store.Users().GetByID(ctx, user.ID)
+		require.NoError(t, err)
+		assert.True(t, updatedUser.IsDisabled)
 
 		apiKey, err := auth.NewAPIKey("ci", "integration", auth.RoleAdmin, "hash", "dagu_123", user.ID)
 		require.NoError(t, err)
 		require.NoError(t, store.APIKeys().Create(ctx, apiKey))
 		require.NoError(t, store.APIKeys().UpdateLastUsed(ctx, apiKey.ID))
+		apiKey.Name = "ci-updated"
+		require.NoError(t, store.APIKeys().Update(ctx, apiKey))
+		updatedAPIKey, err := store.APIKeys().GetByID(ctx, apiKey.ID)
+		require.NoError(t, err)
+		assert.Equal(t, "ci-updated", updatedAPIKey.Name)
 
 		webhook, err := auth.NewWebhook("example", "hash", "dagu_wh_123", user.ID)
 		require.NoError(t, err)
@@ -137,6 +153,11 @@ func TestPostgresControlPlaneStoreIntegration(t *testing.T) {
 		gotWebhook, err := store.Webhooks().GetByDAGName(ctx, "example")
 		require.NoError(t, err)
 		assert.Equal(t, "secret", gotWebhook.HMACSecret)
+		gotWebhook.Enabled = false
+		require.NoError(t, store.Webhooks().Update(ctx, gotWebhook))
+		updatedWebhook, err := store.Webhooks().GetByID(ctx, gotWebhook.ID)
+		require.NoError(t, err)
+		assert.False(t, updatedWebhook.Enabled)
 
 		entry := audit.NewEntry(audit.CategorySystem, "integration_test", user.ID, user.Username)
 		require.NoError(t, store.Audit().Append(ctx, entry))
@@ -160,6 +181,34 @@ func TestPostgresControlPlaneStoreIntegration(t *testing.T) {
 		require.NoError(t, err)
 		require.Len(t, eventResult.Entries, 1)
 		assert.Equal(t, event.ID, eventResult.Entries[0].ID)
+		for i := 0; i < 3; i++ {
+			cursorEvent := &eventstore.Event{
+				ID:            "evt_" + uuid.Must(uuid.NewV7()).String(),
+				OccurredAt:    time.Now().UTC().Add(time.Duration(i) * time.Millisecond),
+				Kind:          eventstore.KindDAGRun,
+				Type:          eventstore.TypeDAGRunUpdated,
+				SourceService: eventstore.SourceServiceServer,
+				UserID:        user.ID,
+				Status:        "blocked",
+			}
+			require.NoError(t, eventSvc.Emit(ctx, cursorEvent))
+		}
+		firstPage, err := eventSvc.Query(ctx, eventstore.QueryFilter{
+			Kind:           eventstore.KindDAGRun,
+			PaginationMode: eventstore.QueryPaginationModeCursor,
+			Limit:          2,
+		})
+		require.NoError(t, err)
+		require.Len(t, firstPage.Entries, 2)
+		require.NotEmpty(t, firstPage.NextCursor)
+		secondPage, err := eventSvc.Query(ctx, eventstore.QueryFilter{
+			Kind:           eventstore.KindDAGRun,
+			PaginationMode: eventstore.QueryPaginationModeCursor,
+			Cursor:         firstPage.NextCursor,
+			Limit:          2,
+		})
+		require.NoError(t, err)
+		require.Len(t, secondPage.Entries, 1)
 
 		sess := &agent.Session{
 			ID:        uuid.Must(uuid.NewV7()).String(),
@@ -182,5 +231,131 @@ func TestPostgresControlPlaneStoreIntegration(t *testing.T) {
 		require.NoError(t, err)
 		require.Len(t, messages, 1)
 		assert.Equal(t, "hello", messages[0].Content)
+		require.NoError(t, store.Sessions().DeleteSession(ctx, sess.ID))
+		require.NoError(t, store.Webhooks().Delete(ctx, gotWebhook.ID))
+		require.NoError(t, store.APIKeys().Delete(ctx, apiKey.ID))
+		require.NoError(t, store.Users().Delete(ctx, user.ID))
+		require.NoError(t, store.Workspaces().Delete(ctx, ws.ID))
+	})
+
+	t.Run("DistributedStores", func(t *testing.T) {
+		attemptKey := exec.GenerateAttemptKey("example", "run-1", "example", "run-1", "attempt-1")
+		owner := exec.CoordinatorEndpoint{ID: "coord-1", Host: "127.0.0.1", Port: 50055}
+		task := &coordinatorv1.Task{
+			Operation:      coordinatorv1.Operation_OPERATION_START,
+			RootDagRunName: "example",
+			RootDagRunId:   "run-1",
+			DagRunId:       "run-1",
+			Target:         "example",
+			QueueName:      "critical",
+			AttemptId:      "attempt-1",
+			AttemptKey:     attemptKey,
+			WorkerSelector: map[string]string{"pool": "blue"},
+		}
+		dispatchStore := store.DispatchTasks()
+		require.NoError(t, dispatchStore.Enqueue(ctx, task))
+		outstanding, err := dispatchStore.CountOutstandingByQueue(ctx, "critical", time.Minute)
+		require.NoError(t, err)
+		assert.Equal(t, 1, outstanding)
+		hasAttempt, err := dispatchStore.HasOutstandingAttempt(ctx, attemptKey, time.Minute)
+		require.NoError(t, err)
+		assert.True(t, hasAttempt)
+
+		claimed, err := dispatchStore.ClaimNext(ctx, exec.DispatchTaskClaim{
+			WorkerID:     "worker-1",
+			PollerID:     "poller-1",
+			Labels:       map[string]string{"pool": "blue"},
+			Owner:        owner,
+			ClaimTimeout: time.Minute,
+		})
+		require.NoError(t, err)
+		require.NotNil(t, claimed)
+		assert.Equal(t, "worker-1", claimed.WorkerID)
+		assert.Equal(t, owner.ID, claimed.Owner.ID)
+		assert.Equal(t, owner.ID, claimed.Task.OwnerCoordinatorId)
+		assert.NotEmpty(t, claimed.ClaimToken)
+
+		gotClaim, err := dispatchStore.GetClaim(ctx, claimed.ClaimToken)
+		require.NoError(t, err)
+		assert.Equal(t, claimed.ClaimToken, gotClaim.ClaimToken)
+		require.NoError(t, dispatchStore.DeleteClaim(ctx, claimed.ClaimToken))
+		_, err = dispatchStore.GetClaim(ctx, claimed.ClaimToken)
+		assert.ErrorIs(t, err, exec.ErrDispatchTaskNotFound)
+
+		heartbeatStore := store.WorkerHeartbeats()
+		require.NoError(t, heartbeatStore.Upsert(ctx, exec.WorkerHeartbeatRecord{
+			WorkerID:        "worker-1",
+			Labels:          map[string]string{"pool": "blue"},
+			LastHeartbeatAt: time.Now().UTC().UnixMilli(),
+		}))
+		heartbeat, err := heartbeatStore.Get(ctx, "worker-1")
+		require.NoError(t, err)
+		assert.Equal(t, "blue", heartbeat.Labels["pool"])
+		heartbeats, err := heartbeatStore.List(ctx)
+		require.NoError(t, err)
+		require.Len(t, heartbeats, 1)
+		deletedStale, err := heartbeatStore.DeleteStale(ctx, time.Now().UTC().Add(time.Hour))
+		require.NoError(t, err)
+		assert.Equal(t, 1, deletedStale)
+
+		leaseStore := store.DAGRunLeases()
+		lease := exec.DAGRunLease{
+			AttemptKey: attemptKey,
+			DAGRun:     exec.NewDAGRunRef("example", "run-1"),
+			Root:       exec.NewDAGRunRef("example", "run-1"),
+			AttemptID:  "attempt-1",
+			QueueName:  "critical",
+			WorkerID:   "worker-1",
+			Owner:      owner,
+		}
+		require.NoError(t, leaseStore.Upsert(ctx, lease))
+		require.NoError(t, leaseStore.Touch(ctx, attemptKey, time.Now().UTC().Add(time.Second)))
+		gotLease, err := leaseStore.Get(ctx, attemptKey)
+		require.NoError(t, err)
+		assert.Equal(t, "worker-1", gotLease.WorkerID)
+		leasesByQueue, err := leaseStore.ListByQueue(ctx, "critical")
+		require.NoError(t, err)
+		require.Len(t, leasesByQueue, 1)
+		allLeases, err := leaseStore.ListAll(ctx)
+		require.NoError(t, err)
+		require.Len(t, allLeases, 1)
+		require.NoError(t, leaseStore.Delete(ctx, attemptKey))
+
+		activeStore := store.ActiveDistributedRuns()
+		active := exec.ActiveDistributedRun{
+			AttemptKey: attemptKey,
+			DAGRun:     exec.NewDAGRunRef("example", "run-1"),
+			Root:       exec.NewDAGRunRef("example", "run-1"),
+			AttemptID:  "attempt-1",
+			WorkerID:   "worker-1",
+			Status:     core.Running,
+		}
+		require.NoError(t, activeStore.Upsert(ctx, active))
+		gotActive, err := activeStore.Get(ctx, attemptKey)
+		require.NoError(t, err)
+		assert.Equal(t, core.Running, gotActive.Status)
+		allActive, err := activeStore.ListAll(ctx)
+		require.NoError(t, err)
+		require.Len(t, allActive, 1)
+		require.NoError(t, activeStore.Delete(ctx, attemptKey))
+	})
+
+	t.Run("ServiceRegistry", func(t *testing.T) {
+		registry := store.Services()
+		require.NoError(t, registry.Register(ctx, exec.ServiceNameCoordinator, exec.HostInfo{
+			ID:     "coord-integration",
+			Host:   "127.0.0.1",
+			Port:   50055,
+			Status: exec.ServiceStatusActive,
+		}))
+		members, err := registry.GetServiceMembers(ctx, exec.ServiceNameCoordinator)
+		require.NoError(t, err)
+		require.Len(t, members, 1)
+		assert.Equal(t, "coord-integration", members[0].ID)
+		require.NoError(t, registry.UpdateStatus(ctx, exec.ServiceNameCoordinator, exec.ServiceStatusInactive))
+		registry.Unregister(ctx)
+		members, err = registry.GetServiceMembers(ctx, exec.ServiceNameCoordinator)
+		require.NoError(t, err)
+		assert.Empty(t, members)
 	})
 }
